style: bit of cleaning up
