+++ backend/src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
-import { AccountsModule } from './accounts/accounts.module';
+import { AccountsModule } from '../accounts/accounts.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
